Move createFragmentFromClipboardData back to core package (#1989)
